requirements.txt
tensorflow_addons
pandas
numpy
matplotlib
jupyter
flask
pytest
pillow